tablex.pairmap: append if key already exists
Previously, a call to tablex.pairmap would overwrite values if a value
was returned for a key that already had a value. Now, when a value is
returned for a key that already exists, we convert the key's value into
a table containing the original value and append the new value being
returned into that table.  Running test-tablex.lua passed after this
modification.