Fix path.normpath
Rewrite path.normpath to first split out path anchor
(drive or UNC path start or '/', '//' or '\') and then iterate
over parts of remaining relative path.

This fixes issues with normpath not working correctly (comparing
to Python's os.path.normpath) in these cases:

* When a path part starts with '..':
    os.path.normpath("A/..B") == "A/..B"
    pl.path.normpath("A/..B") == "B"
* When a path starts with './':
    os.path.normpath("./A") == "A"
    pl.path.normpath("./A") == "./A"
* When a path ends with '/.':
    os.path.normpath("A/.") == "A"
    pl.path.normpath("A/.") == "A/."
* When a path ends with '/':
    os.path.normpath("A/") == "A"
    pl.path.normpath("A/") == "A/"
* When a path starts with '//':
    os.path.normpath("//") == "//"
    pl.path.normpath("//") == "/"

After this commit the only category of mismatches with Python's normpath
is paths starting with '/..': Python collapses that to '/'.

On Windows this commit fixes "C:\.." being normalized to ".".
Fixes #184.