pl.pretty: fix table write with symbol as key
Table keys ordering has been introduce with commit 9aecd3f. Unfortunately
a table sort is apply on all keys, but sort with default comparator is
unable to sort table containing different value types like boolean or table.

A solution is to introduce a custom comparator.
With it keys are now in the following order:
- ordered number
- symbols
- sorted strings
- unorder numbers