chore(docs) Use relative links for local MD files (#341)
This has pros and cons. On the downside it means some places that fetch
just the readme and not other files *and* don't render using the project
base URL might have broken links. On the upside all forks, branches, and
even distributed packages will link to their own respective local files
and versions.