docs: Drop last updated timestamp so builds are deterministic
The documentation is already versioned with the edition of Penlight.
Having a timestamp is counter productive. It does not allow you to
deduce what version of the source was actually built from because it is
the build date, not the commit info. It also makes builds
non-deterministic which makes it hard for contributors adds noise to the
history.