chore(build): Refactor rockspec to dodge luarocks 3.1.3 bug
Setting source.branch works for most cases, but when the branch is
actually a tag and not a branch then a Luarocks 3.1.3 bug rears
and tries to concatenate a nil. This dodges that bullet by setting
source.tag explicitly to not trigger the branch_or_tag concatenation.

All the other changes are cosmetic and for readability/maintainability.