Updates tablex readonly to assert on non-iterable/indexable data
Consider the case where non-indexable/non-iterable data is provided to readonly. We accomplish this by creating an empty table and using metatable manipulation. We make the original data the index of the empty table, attempt to use pairs/ipairs on the original data, and then we obfuscate the metatable to restrict access to the underlying data. However if we have non-indexable or non-iterable data like a number or a string, we cannot access the data at all. ipairs/pairs will not give us anything fruitful and attempting to poke the value directly via key/index will result in nothing because the underlying data is not indexable/iterable. The only way to access the data is via debug where the obfuscated metatable can be obtained, and the readonly property invalidated by accessing the underlying table via the __index property.

Adding this assert prevents this invalid and intuitive use of readonly that people coming from languages that interface with Lua (C/C++) might confuse with something like const.